api/requirements.txt
fastapi==0.47.1
uvicorn==0.11.2
pydantic
PyJWT
bcrypt